backend/requirements.txt
fastapi
uvicorn
numpy
opencv-python
mediapipe
openai
python-dotenv